requirements.man.txt
# Django Framework
Django==3.0.7

# Rest Framework
djangorestframework==3.11.0

# Swagger
drf-yasg==1.17.1

# PostgreSQL driver
psycopg2-binary==2.8.5

# Read environment vars
python-dotenv==0.13.0

# coding conventions
autopep8==1.5.3

# WSGI config
gunicorn==20.0.4

# database connection as url
dj-database-url==0.5.0

# Manage static file
whitenoise==5.1.0

# extensions: main purpose == gen diagrams
django-extensions==2.2.9
# this shit keep getting error
# pygraphviz==1.5
# --install-option="--include-path=/usr/local/include/" --install-option="--library-path=/usr/local/lib/" pygraphviz

# OTP 6 digits numb
drfpasswordless==1.5.4

# Django CORS header
django-cors-headers==3.4.0

# Django Filter
django-filter==2.3.0

# Django GG Drive || Cloud storage maybe ?
django-googledrive-storage==1.5.2

# Django Rest Password Reset
django-rest-passwordreset==1.1.0